Move Plane/Sphere helper classes to math/ folder
