+++ CMakeLists.txt
@@ -152,7 +152,7 @@ list(APPEND SRC_FOLDERS
 	src/graphics
 	src/graphics/dummy
 	src/graphics/opengl
-	src/gui
+	src/math
 	src/pigui
 	src/scenegraph
 	src/ship
